move nasty radix themes js code into js files
